libdrgn: bring back dwfl_core_file_report() bug workaround
This workaround was originally present in commit e5874ad18a53 ("libdrgn:
use libdwfl"). We dropped in in commit 6a13d74c0c45 ("libdrgn: build
with bundled elfutils") because the bundled version of elfutils had the
fix. We forgot to bring it back in commit 4c5c5f384232 ("Remove bundled
version of elfutils") even though we support versions without the fix.

Reported-by: Serapheim Dimitropoulos <serapheim@delphix.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>