Live userspace process support for the Threads API
PR #129 introduced the Threads API to drgn, but did not include support
for live userspace processes. This commit changes that, both adding
support for the existing Threads API, as well as extending it with a
few additional methods which only make sense in the context of a
live process, most notably `Thread.pause()` and `Thread.resume()`

At present, only x86 is supported, but the plan is to support other
architectures in the near future.

Signed-off-by: Kevin Svetlitski <svetlitski@fb.com>