cli: diagnose when someone tries to run a binary as a script
This is a common mistake:

  $ drgn core_dump
  Traceback (most recent call last):
    File "/usr/bin/drgn", line 33, in <module>
      sys.exit(load_entry_point('drgn==0.0.16', 'console_scripts', 'drgn')())
    File "/usr/lib/python3.10/site-packages/drgn/internal/cli.py", line 133, in main
      runpy.run_path(args.script[0], init_globals=init_globals, run_name="__main__")
    File "/usr/lib/python3.10/runpy.py", line 268, in run_path
      code, fname = _get_code_from_file(run_name, path_name)
    File "/usr/lib/python3.10/runpy.py", line 242, in _get_code_from_file
      code = compile(f.read(), fname, 'exec')
  ValueError: source code string cannot contain null bytes

The user intends to debug the core dump, but they've actually specified
the core dump as a Python script to run. The error message from the
runpy internals does not make that clear. So, let's catch this earlier
by doing a quick-and-dirty test of the file magic to see if it looks
like a core dump or other ELF file. If so, we exit with a more helpful
message:

  $ drgn core_dump
  error: core_dump is a core dump
  Did you mean "-c core_dump"?
  $ drgn /usr/bin/ls
  error: /usr/bin/ls is a binary, not a drgn script

Signed-off-by: Omar Sandoval <osandov@osandov.com>