Add gen_strswitch.py
We have multiple places where we match an input string against several
cases:

* drgn_lexer_c() checks C identifiers against a runtime hash table of C
  keywords.
* linux_kernel_object_find() has an if-else ladder of checks for object
  names.
* drgn_debug_info_find_sections() loops over an array of ELF section
  names to look for sections we need.
* libdrgn/build-aux/gen_arch.awk generates a compile-time trie using
  nested switch statements to match register names.

This commit adds a script, gen_strswitch.py, that can hopefully be used
to replace all of these. gen_strswitch.py generalizes the compile-time
trie idea from gen_arch.awk in a few ways:

* It has syntax and semantics based on C switch statements.
* It supports both null-terminated strings and strings with an explicit
  length.
* It compresses unique substrings to calls to strcmp(), strncmp(), or
  memcmp() when appropriate.

In benchmarks, this approach is more performant than the above options
as well as a candidate based on gperf, while resulting in machine code
around the same size as the straightforward if-else ladder approach.

Future commits will convert the use cases above to use this script.

Signed-off-by: Omar Sandoval <osandov@osandov.com>