Introduce API for debugging a program without running it
Debuggers like `gdb` allow the user to specify a particular binary
to debug (and potentially arguments to be used) without immediately
starting a new process. Such functionality is often convenient, whether
it be for setting breakpoints before the program starts, or querying the
various types in the binary without ever running it at all.

This commit introduces such functionality to `drgn` in the form of the
new APIs `drgn_program_set_argv`, `drgn_program_from_argv`, and
`drgn_program_run`. The first two allow the user the set the target to a
particular binary without running it, as well as specifying command
line arguments that will be passed when a new process is created to run
the given binary. `drgn_program_run` then starts a new process with the
executable and arguments passed to the former 2 functions.

Signed-off-by: Kevin Svetlitski <svetlitski@fb.com>